Prefix/Suffix Support, Bug Fixes
Added an option to add a prefix and suffix (-p, -s) to your overflow string. For example, it is now possible to run
USER AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA\r\n
without modifying the source code.

There was a bug that occurred when sending the bad character list if there wasn't enough space before the offset to fit the list of bad characters. This bug was squashed.

Made the script only accept y/n to launch the payload. It no longer defaults to "launch."